requirements.txt
pip
requests
pytest
logger
gitpython